BUG FIX: Player now correctly choses next platform.
zRot is in degrees and i was using it as if in radians. zRot is now zRotDeg.

Also, somehow the draw code ended up getting changed, but for the good. the draw function signatures now directly reference the Verts class and TexCoords class. It was lazy of me not to make them do that in the first place.